1 - Add leave workspace feature

+++ src/server/api/routers/workspace.ts
@@ -351,6 +351,27 @@ export const workspaceRouter = router({
       })
     )
     .mutation(async ({ ctx, input }) => {
+      // throw error if there will be no researcher admin left when leaving workspace
+      const workspace = await ctx.prisma.workspace.findUnique({
+        where: {
+          id: input.workspaceId,
+        },
+        include: {
+          workspace_user: true,
+        },
+      });
+
+      const researcherAdmins = workspace?.workspace_user.filter(
+        (user) => user.workspace_role === "Researcher Admin"
+      );
+
+      if (researcherAdmins?.length === 1) {
+        throw new TRPCError({
+          code: "INTERNAL_SERVER_ERROR",
+          message: "Workspace must have at least one researcher admin.",
+        });
+      }
+
       const workspaceUser = await ctx.prisma.workspace_user.delete({
         where: {
           workspaceid_userid: {
